refactor: question 타입, 상수 분리
